flask_backend/requirements.txt
flask==3.1.2
flask-cors==6.0.1
flask-sqlalchemy==3.1.1
flask-migrate==4.1.0
requests==2.32.5
python-dotenv==1.1.1
alembic==1.16.5